fix bug in ConfigParams15 deserialization; add toKeyPair(String) to create key pair from mnemonic string, where words separated with space; change tweetnacl java artifact's name; remove MsgUtils.createInternalMessage without source address (replaced with createInternalMessage with source address and multi-currency); replace Message with MessageRelaxed where possible; improve extra-currency serialization; back to java 8;

+++ tlb/src/main/java/org/ton/java/tlb/CurrencyCollection.java
@@ -32,7 +32,7 @@ public class CurrencyCollection implements Serializable {
     if (nonNull(extraCurrencies)) {
       dict =
           extraCurrencies.serialize(
-              k -> CellBuilder.beginCell().storeUint((BigInteger) k, 32).endCell().getBits(),
+              k -> CellBuilder.beginCell().storeUint((Long) k, 32).endCell().getBits(),
               v -> CellBuilder.beginCell().storeVarUint((BigInteger) v, 5).endCell());
     }
     return CellBuilder.beginCell().storeCoins(coins).storeDict(dict).endCell();
@@ -44,7 +44,7 @@ public class CurrencyCollection implements Serializable {
         .extraCurrencies(
             cs.loadDictE(
                 32,
-                k -> k.readUint(32),
+                k -> k.readUint(32).longValue(),
                 v -> CellSlice.beginParse(v).loadVarUInteger(BigInteger.valueOf(5))))
         .build();
   }
